(fix) (UMCS-426) changes from review.
